refactors: lowercase models
