+++ src/components/orchestrator/SurveyContainer.tsx
@@ -6,8 +6,8 @@ import Button from '@codegouvfr/react-dsfr/Button'
 import { MODE_TYPE } from '@/constants/mode'
 import { PAGE_TYPE } from '@/constants/page'
 import { declareComponentKeys, useTranslation } from '@/i18n'
-import type { InternalPageType } from '@/models/Page'
 import type { LunaticOverview } from '@/models/lunaticType'
+import type { InternalPageType } from '@/models/page'
 
 import type { OrchestratorProps } from './Orchestrator'
 import { SequenceHeader } from './SequenceHeader'
