+++ src/components/orchestrator/customPages/EndPage.tsx
@@ -2,7 +2,7 @@ import { fr } from '@codegouvfr/react-dsfr'
 
 import { useDocumentTitle } from '@/hooks/useDocumentTitle'
 import { declareComponentKeys, useTranslation } from '@/i18n'
-import type { StateData } from '@/models/StateData'
+import type { StateData } from '@/models/stateData'
 
 /**
  * Display time at which user has sent its answers.
